pyproject.toml
[project]
name = "opp-uruguay-pipeline"
version = "0.1.0"
description = "End-to-end data pipeline for Uruguay OPP national budget data"
requires-python = ">=3.11"
dependencies = [
    "polars==1.24.0",
    "httpx==0.28.1",
    "beautifulsoup4==4.13.3",
    "google-cloud-storage==2.19.0",
    "google-cloud-bigquery==3.27.0",
    "pdfplumber==0.11.6",
    "pyarrow==18.1.0",
    "lxml==5.3.1",
]

[project.optional-dependencies]
dev = [
    "pip-audit==2.8.0",
    "ruff==0.9.7",
]
dbt = [
    "dbt-bigquery==1.9.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipelines"]

[tool.ruff]
line-length = 100
target-version = "py311"
